Complete exercise 03 of flexbox
In the ul selector I have written margin: 0 auto 0 0; this sets an
auto margin on the right that expands ocupying all the aviable space.

In order to acomplish this task we could also have written
justify-content: space between; (on .header) and then grup the two
right elements of the header inside a div.