Complete exercise 05 of flexbox
In order to acomplish the task of spacing the elements on
the main-message, instead of adding margins we also could have made
something like adding dsiplay flex into .main-message, change the
flex-direction to column and adding a gap. But this aproach had the
prblem that we needed different margins on the bottom and the top.